sliced up the components further and refactored all the functions to hooks with useCallbacks
